EditorPaneSchemaSidebar
